requirements.txt
aiogram>=3.0.0
qrcode[pil]>=7.0
python-dotenv>=1.0.0
pillow>=10.0.0